% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pokemon.R
\docType{data}
\name{Pokemon}
\alias{Pokemon}
\title{Pokemon Data Visualized}
\format{
A data frame with 800 observations with 13 columns.
\describe{
   \item{Index}{The Pokedex number of the Pokemon}
   \item{Name}{The name of the Pokemon}
   \item{Primary_Type}{The primary type of the Pokemon}
   \item{Secondary_Type}{The secondary type of the Pokemon}
   \item{Total}{The sum of the Pokemon's Base Stats}
   \item{HP}{The Pokemon's HP stat}
   \item{Attack}{The Pokemon's Attack stat}
   \item{Defense}{The Pokemon's Defense stat}
   \item{Sp_Atk}{The Pokemon's Special Attack stat}
   \item{Sp_Def}{The Pokemon's Special Defense stat}
   \item{Speed}{The Pokemon's Speed stat}
   \item{Generation}{The generation the Pokemon was introduced}
   \item{Legendary}{Shows whether the Pokemon is considered a Legendary/Mythical (True) or not (False)}
}
}
\usage{
Pokemon
}
\description{
This dataset contains information about the Pokemon introduced in the first
6 generations of the franchise. It has information about their Pokedex
number, type, stats, and to determine whether they are a legendary/mythical
or not.
}
\keyword{datasets}
